dev-13777 edit counterpart modal in its own file
